requirements.txt
Flask==2.0.1
h5py==3.1.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
protobuf==3.14.0
requests==2.25.1
requests-oauthlib==1.3.0
urllib3==1.26.5
Werkzeug==2.0.1
gunicorn==19.8.0
PyGithub==1.55
python-dotenv